refactor(vmm): replace static CPU templates for x86_64
Replace the static CPU templates introduced in PR #3076 with ones that
are introduced in PR #3598, are compatibile with main branch and are
defined using `CustomCpuTemplate`.
