--- src/vmm/src/guest_config/x86_64/static_cpu_templates_new/mod.rs
@@ -1,13 +1,0 @@
-// Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-/// Module with C3 CPU template for x86_64
-pub mod c3;
-/// Module with T2 CPU template for x86_64
-pub mod t2;
-/// Module with T2A CPU template for x86_64
-pub mod t2a;
-/// Module with T2CL CPU template for x86_64
-pub mod t2cl;
-/// Module with T2S CPU template for x86_64
-pub mod t2s;
